Introduced ContentTypeVisitorTrait
