+++ libraries/incubator/Renderer/HtmlRenderer.php
@@ -16,6 +16,7 @@ use Joomla\Content\Type\Attribution;
 use Joomla\Content\Type\Columns;
 use Joomla\Content\Type\Compound;
 use Joomla\Content\Type\DefaultMenu;
+use Joomla\Content\Type\Dump;
 use Joomla\Content\Type\Headline;
 use Joomla\Content\Type\HorizontalLine;
 use Joomla\Content\Type\Icon;
@@ -268,11 +269,11 @@ class HtmlRenderer extends Renderer
 	/**
 	 * Dump an item
 	 *
-	 * @param   ContentTypeInterface $dump The dump
+	 * @param   Dump $dump The dump
 	 *
 	 * @return  void
 	 */
-	public function visitDump(ContentTypeInterface $dump)
+	public function visitDump(Dump $dump)
 	{
 		$this->write('<pre>' . $this->dumpEntity($dump->item) . '</pre>');
 	}
